+++ src/advancedsettings.js
@@ -153,6 +153,32 @@ class AdvancedSettings {
     }
     return this._version.updateResource(URIs.REQUEST_BODY, [], payload);
   }
+
+  getEvasivePathMatch() {
+    if (this._options.policy) {
+      return this._policyProvider.policyId().then(policyId => {
+        return this._version.readResource(URIs.SECURITY_POLICY_EVASIVE_PATH_MATCH, [policyId]);
+      });
+    }
+    return this._version.readResource(URIs.EVASIVE_PATH_MATCH, []);
+  }
+
+  updateEvasivePathMatch(enable) {
+    let data = JSON.parse(
+      fs.readFileSync(__dirname + '/../templates/evasivepathmatch.json', 'utf8')
+    );
+    data.enablePathMatch = enable;
+    if (this._options.policy) {
+      return this._policyProvider.policyId().then(policyId => {
+        return this._version.updateResource(
+          URIs.SECURITY_POLICY_EVASIVE_PATH_MATCH,
+          [policyId],
+          data
+        );
+      });
+    }
+    return this._version.updateResource(URIs.EVASIVE_PATH_MATCH, [], data);
+  }
 }
 
 module.exports = {